src/main/java/nl/novi/eindopdrachtBackenSystemGoldencarrot/utilsGeneralMethods/BindingValidator.java
package nl.novi.eindopdrachtBackenSystemGoldencarrot.utilsGeneralMethods;

import org.springframework.stereotype.Component;
import org.springframework.validation.BindingResult;
import org.springframework.validation.FieldError;

@Component
public class BindingValidator {

    public static String validateInput(BindingResult br) {
        if (br.hasFieldErrors()) {
            StringBuilder sb = new StringBuilder();
            for (FieldError fe : br.getFieldErrors()) {
                sb.append(fe.getField() + ": ");
                sb.append(fe.getDefaultMessage() + "\n");
            }
            return sb.toString();
        } else return null;
    }
}
